TINY-9523: moved tooglableIconButton in a new file in view dir
